add auth

+++ resources/views/cultivation/index.blade.php
@@ -16,29 +16,36 @@
         @endif
         <div class="box">
             <div class="box-header" >
-                {{Form::open(array('url' => 'cultivation/create', 'method' => 'GET','style'=>'width:80%;margin:auto; text-align:center'))}}
-                    {{ Form::submit('Tambah Budidaya Baru', array('class' => 'btn btn-info btn-flat')) }}
+                @if(Auth()->user()->role_id == 1)
+                {{Form::open(['url' => 'cultivation/create', 'method' => 'GET','style'=>'width:80%;margin:auto; text-align:center'])}}
+                    {{ Form::submit('Tambah Budidaya Baru', ['class' => 'btn btn-info btn-flat']) }}
                 {{Form::close()}}
-            </div>
 
-            {{Form::open(array('url' => 'cultivation/search', 'method' => 'GET'))}}
-            <div class="input-group " style="width: 80%; margin: auto;">
+                @endif
+
+
+                    {{Form::open(array('url' => 'cultivation/', 'method' => 'GET'))}}
+                    <div class="input-group " style="width: 80%; margin: auto;">
                 <span class="input-group-addon">
                         {{ Form::label('search','Name Search')}}
                 </span>
-                {{ Form::text('search','',['class' => 'form-control'])}}
-                <span class="input-group-btn">
+                        {{ Form::text('search','',['class' => 'form-control'])}}
+                        <span class="input-group-btn">
                         {{ Form::submit('search', array('class' => 'btn btn-info btn-flat')) }}
                 </span>
+                    </div>
+                    {{Form::close()}}
             </div>
-            {{Form::close()}}
             <!-- /.box-header -->
             <div class="box-body">
                 <table id="example1" class="table table-bordered table-striped">
                     <thead>
                     <tr>
                         <th>Jenis Budidaya</th>
-                        <th>Action</th>
+                        @if(Auth()->user()->role_id == 1)
+                            <th>Action</th>
+                        @endif
+
                     </tr>
                     </thead>
                     <tbody>
@@ -46,23 +53,26 @@
                         <tr>
                             <td>{{$d->description}}</td>
                             </td>
-                            <td>
-                                <div class="input-group input-group-sm">
-                                    <span class="input-group-btn">
-                                        {{Form::open(array('url' => 'cultivation/'.$d->id.'/edit', 'method' => 'GET'))}}
-                                            {{ Form::submit('Edit', array('class' => 'btn btn-warning btn-flat')) }}
-                                        {{Form::close()}}
-                                    </span>
 
-                                    <span class="input-group-btn">
-                                        {{Form::open(array('url' => 'cultivation/'.$d->id, 'method' => 'delete'))}}
+                            @if(Auth()->user()->role_id == 1)
+                                <td>
+                                    <div class="input-group input-group-sm">
+                                        <span class="input-group-btn">
+                                            {{Form::open(array('url' => 'cultivation/'.$d->id.'/edit', 'method' => 'GET'))}}
+                                                {{ Form::submit('Edit', array('class' => 'btn btn-warning btn-flat')) }}
+                                            {{Form::close()}}
+                                        </span>
 
-                                            {{ Form::submit('Delete', array('class' => 'btn btn-danger btn-flat')) }}
-                                        {{Form::close()}}
-                                    </span>
-                                </div>
+                                        <span class="input-group-btn">
+                                            {{Form::open(array('url' => 'cultivation/'.$d->id, 'method' => 'delete'))}}
 
-                            </td>
+                                                {{ Form::submit('Delete', array('class' => 'btn btn-danger btn-flat')) }}
+                                            {{Form::close()}}
+                                        </span>
+                                    </div>
+
+                                </td>
+                            @endif
                         </tr>
                     @endforeach
                     </tbody>
